+++ src/App.jsx
@@ -1,11 +1,11 @@
-import AppHeader from './components/app-header/AppHeader'
+import Header from './components/common/header/Header'
 import Constructor from './components/pages/constructor/Constructor'
 import Modal from './components/common/modal/Modal'
 
 function App() {
     return (
         <>
-            <AppHeader />
+            <Header />
 
             <main>
                 <Constructor />
