+++ src/components/common/total-price/TotalPrice.module.css
@@ -1,0 +1,10 @@
+.total_price {
+    display: flex;
+    align-items: center;
+    gap: 0.55em;
+}
+
+.total_price svg {
+    width: 33px;
+    height: 33px;
+}
