+++ src/components/constructor-page-components/burger-constructor-bun/BurgerConstructorBun.jsx
@@ -1,0 +1,36 @@
+import PropTypes from 'prop-types'
+
+import { useSelector } from 'react-redux'
+
+import { ConstructorElement } from '@ya.praktikum/react-developer-burger-ui-components'
+
+import styles from './BurgerConstructorBun.module.css'
+
+export default function BurgerConstructorBun({ position }) {
+    const bun = useSelector(store => store.constructorIngredientList.bun)
+
+    const classForEmpty = `constructor-element constructor-element_pos_${position}
+        ${styles[`bun_${position}`]}
+        ${styles.constructor_element}`
+
+    const positionText = {
+        top: 'верх',
+        bottom: 'низ',
+    }
+
+    return bun._id ? (
+        <ConstructorElement
+            type={position}
+            isLocked={true}
+            text={`${bun.name} (${positionText[position]})`}
+            price={bun.price}
+            thumbnail={bun.image}
+        />
+    ) : (
+        <div className={classForEmpty}>Выберите булки</div>
+    )
+}
+
+BurgerConstructorBun.propTypes = {
+    position: PropTypes.oneOf(['top', 'bottom'])
+}
