+++ src/components/constructor-page-components/ingredient-item/IngredientItem.module.css
@@ -1,0 +1,24 @@
+.ingredient_item {
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 8px;
+    cursor: pointer;
+}
+
+.image {
+    width: 240px;
+    height: 120px;
+}
+
+.price {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 8px;
+}
+
+.name {
+    text-align: center;
+}
