--- src/components/pages/constructor/burger-constructor/BurgerConstructor.module.css
@@ -1,70 +1,0 @@
-.ingredients {
-    position: relative;
-}
-
-.empty_constructor {
-    flex-grow: 1 !important;
-    margin: 0 !important;
-}
-
-.ingredients_block {
-    position: absolute;
-    top: 100px;
-    display: flex;
-    width: calc(100% - 20px);
-    height: 560px;
-    flex-direction: column;
-    align-items: flex-end;
-    gap: 16px;
-}
-
-.ingredient_list {
-    display: flex;
-    flex-direction: column;
-    position: relative;
-    gap: 16px;
-    list-style-type: none;
-    margin: 0 -20px 0 0;
-    padding: 0;
-    overflow-y: auto;
-}
-
-.ingredient_item {
-    margin-right: 8px;
-    flex-grow: unset;
-    display: flex;
-    align-items: center;
-    gap: 7px;
-}
-
-.bun_top,
-.bun_bottom {
-    flex-grow: unset;
-}
-
-.constructor_element {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.order_block {
-    position: absolute;
-    bottom: 50px;
-    right: 16px;
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-    gap: 40px;
-}
-
-.total_price {
-    display: flex;
-    align-items: center;
-    gap: 0.55em;
-}
-
-.total_price svg {
-    width: 33px;
-    height: 33px;
-}
