--- src/components/pages/constructor/burger-ingredients/BurgerIngredients.module.css
@@ -1,45 +1,0 @@
-.ingredient_elements {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 32px 24px;
-    list-style-type: none;
-    margin: 0;
-}
-
-.burger_ingredients {
-    display: grid;
-    grid-template-rows: min-content 1fr;
-    gap: 40px;
-    overflow: hidden;
-}
-
-.ingredients_inner {
-    position: relative;
-    overflow-y: auto;
-    scroll-behavior: smooth;
-}
-
-.list_item {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 8px;
-    cursor: pointer;
-}
-
-.image {
-    width: 240px;
-    height: 120px;
-}
-
-.price {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 8px;
-}
-
-.name {
-    text-align: center;
-}
